# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: LLMask API
  version: "0.1.0"
  description: |
    **LLMask** is an OpenAI-compatible privacy proxy that detects and masks sensitive data
    (PII, secrets, credentials) before forwarding requests to any LLM provider, then
    remaps pseudonyms back in the response.

    ## Authentication
    When `ADMIN_KEY` is set in the environment, all `/dashboard/api/*` routes require the
    key to be supplied either as:
    - Query parameter: `?key=<admin-key>`
    - Header: `X-LLMask-Key: <admin-key>`

    Proxy routes (`/v1/*`) have no built-in authentication in the community edition.

  contact:
    name: LLMask OSS
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /v1/chat/completions:
    post:
      tags: [Proxy]
      summary: Chat completions proxy (OpenAI-compatible)
      description: |
        Drop-in replacement for the OpenAI `/v1/chat/completions` endpoint.
        LLMask detects PII and secrets in the request, replaces them with pseudonyms,
        forwards the sanitised request to the configured provider, then remaps
        pseudonyms back in the response.

        Streaming (`"stream": true`) is supported and returns Server-Sent Events in the
        standard OpenAI format.
      operationId: proxyChatCompletions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatCompletionsRequest"
            example:
              model: gpt-4o
              messages:
                - role: user
                  content: "Summarise the email from john.doe@acme.com about project Falcon"
              stream: false
      responses:
        "200":
          description: Chat completion (non-streaming) or SSE stream
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatCompletionsResponse"
            text/event-stream:
              schema:
                type: string
                description: Server-Sent Events stream in OpenAI delta format
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/RateLimited"
        "502":
          $ref: "#/components/responses/BadGateway"
  /v1/messages:
    post:
      tags: [Proxy]
      summary: Anthropic Messages API proxy
      description: |
        Drop-in replacement for the Anthropic `/v1/messages` endpoint.
        Accepts the Anthropic Messages request format; LLMask masks PII before
        forwarding to the configured provider.
      operationId: proxyMessages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnthropicMessagesRequest"
            example:
              model: claude-sonnet-4-20250514
              max_tokens: 1024
              messages:
                - role: user
                  content: "Analyse the report for Alice Smith at alice@example.com"
      responses:
        "200":
          description: Anthropic messages response or SSE stream
          content:
            application/json:
              schema:
                type: object
                description: Anthropic Messages API response (passthrough)
            text/event-stream:
              schema:
                type: string
                description: Anthropic SSE stream
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/BadGateway"
  /v1/responses:
    post:
      tags: [Proxy]
      summary: OpenAI Responses API proxy
      description: |
        Drop-in replacement for the OpenAI `/v1/responses` endpoint.
        Uses the structured `input` array format introduced in the Responses API.
      operationId: proxyResponses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResponsesRequest"
            example:
              model: gpt-4o
              input:
                - type: message
                  role: user
                  content:
                    - type: input_text
                      text: "Summarise the data for SSN 123-45-6789"
              stream: true
      responses:
        "200":
          description: Responses API result or SSE stream
          content:
            application/json:
              schema:
                type: object
                description: OpenAI Responses API response (passthrough)
            text/event-stream:
              schema:
                type: string
                description: OpenAI Responses API SSE stream
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/BadGateway"

  # ── Files ─────────────────────────────────────────────────────────────────
  /v1/files/anonymize:
    post:
      tags: [Files]
      summary: Anonymize a file
      description: |
        Upload a file; LLMask extracts its text, detects PII/secrets, and returns
        anonymised content.

        **Supported file types:**
        - Text: `.txt`, `.md`, `.json`, `.csv`, `.xml`, `.html`
        - Documents: `.pdf`, `.docx`, `.xlsx`, `.pptx`
        - Images: `.jpg`, `.jpeg`, `.png`, `.webp` (EXIF metadata is stripped)

        Maximum file size: **10 MB**.
      operationId: anonymizeFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: The file to anonymise
                scope_id:
                  type: string
                  description: |
                    Scope identifier for this mapping session. If omitted a random UUID
                    is generated. Use the same scope_id across requests to share the
                    pseudonym dictionary.
      responses:
        "200":
          description: File anonymised successfully
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/AnonymizeTextResponse"
                  - $ref: "#/components/schemas/AnonymizeImageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          description: File blocked by LLMask policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "422":
          description: Failed to extract text from the document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Dashboard – Config ────────────────────────────────────────────────────
...
components:
  securitySchemes:
    adminKey:
      type: apiKey
      in: header
      name: X-LLMask-Key
      description: |
        Admin key for dashboard API access. Set via `ADMIN_KEY` environment variable.
        Can also be passed as query parameter `?key=<admin-key>`.
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    BadGateway:
      description: Upstream LLM provider error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
            type:
              type: string
            code:
              type: string
            param:
              type: string
              nullable: true

    # ── Health ────────────────────────────────────────────────────────────────
...
    ChatMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [system, user, assistant, tool, function, developer]
        content:
          type: string
          nullable: true
        name:
          type: string
        tool_calls:
          type: array
          items:
            type: object
        tool_call_id:
          type: string
    ChatCompletionsRequest:
      type: object
      required: [model, messages]
      properties:
        model:
          type: string
          example: gpt-4o
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
        temperature:
          type: number
          minimum: 0
          maximum: 2
          example: 0.7
        max_tokens:
          type: integer
          example: 1024
        top_p:
          type: number
          example: 1
        stream:
          type: boolean
          default: false
        stop:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        tools:
          type: array
          items:
            type: object
        tool_choice:
          type: object
    ChatCompletionsResponse:
      type: object
      properties:
        id:
          type: string
          example: chatcmpl-abc123
        object:
          type: string
          enum: [chat.completion]
        created:
          type: integer
          description: Unix timestamp
        model:
          type: string
        choices:
          type: array
          items:
            type: object
            properties:
              index:
                type: integer
              message:
                $ref: "#/components/schemas/ChatMessage"
              finish_reason:
                type: string
                nullable: true
        usage:
          type: object
          properties:
            prompt_tokens:
              type: integer
            completion_tokens:
              type: integer
            total_tokens:
              type: integer

    # ── Proxy: Anthropic Messages ─────────────────────────────────────────────
    AnthropicMessagesRequest:
      type: object
      required: [model, max_tokens, messages]
      properties:
        model:
          type: string
          example: claude-sonnet-4-20250514
        max_tokens:
          type: integer
          example: 1024
        messages:
          type: array
          items:
            type: object
            required: [role, content]
            properties:
              role:
                type: string
                enum: [user, assistant]
              content:
                oneOf:
                  - type: string
                  - type: array
                    items:
                      type: object
        system:
          type: string
        temperature:
          type: number
        stream:
          type: boolean
          default: false

    # ── Proxy: Responses API ──────────────────────────────────────────────────
    ResponsesRequest:
      type: object
      required: [model, input]
      properties:
        model:
          type: string
          example: gpt-4o
        input:
          description: Array of message objects in Responses API format
          oneOf:
            - type: string
            - type: array
              items:
                type: object
        instructions:
          type: string
          description: System-level instructions
        temperature:
          type: number
        max_output_tokens:
          type: integer
        stream:
          type: boolean
          default: false
        tools:
          type: array
          items:
            type: object

    # ── Files ─────────────────────────────────────────────────────────────────
    AnonymizeTextResponse:
      type: object
      properties:
        status:
          type: string
          enum: [ok]
        fileName:
          type: string
          example: report.pdf
        scopeId:
          type: string
          example: "550e8400-e29b-41d4-a716-446655440000"
        type:
          type: string
          enum: [text, document]
        transformedCount:
          type: integer
          description: Number of entities that were anonymised
        originalSize:
          type: integer
          description: Original file size in bytes
        anonymizedContent:
          type: string
          description: Anonymised plain text content
        format:
          type: string
          description: Detected document format (for documents only)
        pageCount:
          type: integer
          description: Number of pages (for documents only)
        documentMetadata:
          type: object
          description: Extracted document metadata (for documents only)
    AnonymizeImageResponse:
      type: object
      properties:
        status:
          type: string
          enum: [ok]
        fileName:
          type: string
          example: photo.jpg
        scopeId:
          type: string
        type:
          type: string
          enum: [image]
        format:
          type: string
          example: jpeg
        originalSize:
          type: integer
        sanitizedSize:
          type: integer
        strippedMetadataChunks:
          type: integer
          description: Number of EXIF/metadata chunks removed
        metadataFound:
          type: array
          items:
            type: string
          description: Names of metadata types that were found and stripped
        sanitizedBase64:
          type: string
          format: byte
          description: Base64-encoded image with metadata removed

    # ── Dashboard – Config ────────────────────────────────────────────────────
...